Bug 801383 Fix fallout from bug 779473 - make nsresult an enum class. Fix nsMsgPrompts and get mailnews/compose compiling on Mac. Patch originally by Ian Neil, updated by Standard8. r=Neil. CLOSED TREE

+++ mailnews/compose/src/nsMsgAppleDouble.h
@@ -34,7 +34,7 @@
 								/* 	End of a Part.				*/
 
 					
-#define errFileOpen		NS_MSG_UNABLE_TO_OPEN_TMP_FILE
+#define errFileOpen		static_cast<uint32_t>(NS_MSG_UNABLE_TO_OPEN_TMP_FILE)
 #define errFileWrite	-202 /*Error writing temporary file.*/
 #define errUsrCancel	-2  /*MK_INTERRUPTED */
 #define errDecoding		-1
